requirements.txt
absl-py==2.1.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arm_pytorch_utilities==0.4.3
array_record==0.6.0
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1734292804024/work
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269734548/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
distrax==0.1.5
-e git+ssh://git@github.com/kvablack/dlimp.git@5edaa4691567873d495633f2708982b42edf1972#egg=dlimp
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.2.10
flax==0.7.5
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
gdown==5.2.0
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humanize==4.12.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1740856895/work
isoduration==20.11.0
jax==0.4.20
jaxlib==0.4.20+cuda11.cudnn86
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
-e git+ssh://git@github.com/haosulab/ManiSkill.git@91e139683efddf1ffe6d0caa354d1c3b5ced5731#egg=mani_skill
-e git+https://github.com/simpler-env/SimplerEnv.git@052eaea54dc7c37447cf40c499134b55534bbde4#egg=mani_skill2_real2sim&subdirectory=ManiSkill2_real2sim-cd45dd27dc6bb26d048cb6570cdab4e3f935cc37
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mediapy==1.2.0
mistune==3.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
mplib==0.1.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvcc-cu12==12.8.61
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu11==9.7.1.26
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.41
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.7.53
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
-e git+ssh://git@github.com/octo-models/octo.git@653c54acde686fde619855f2eac0dd6edad7116b#egg=octo
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.4.8
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
proglog==0.1.10
prometheus_client==0.21.1
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==3.20.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663128538/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.1.3
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pynvml==12.0.0
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-json-logger==3.2.1
pytorch-kinematics==0.7.5
pytorch-seed==0.2.0
PyVirtualDisplay==3.0
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1738270958168/work
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
Rtree==1.3.0
ruckig==0.14.0
safetensors==0.5.3
sapien==3.0.0b1
scipy==1.12.0
Send2Trash==1.8.3
shtab==1.7.1
simple-parsing==0.1.7
simplejson==3.20.1
-e git+ssh://git@github.com/simpler-env/SimplerEnv.git@52a5088ca4bfc3a7159828af08874a198d6f95c3#egg=simpler_env
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.7
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.23.0
tensorstore==0.1.72
termcolor==2.5.0
terminado==0.18.1
tf-agents==0.19.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
toppra==0.6.0
torch==2.3.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615898999/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.49.0
transforms3d==0.4.2
trimesh==4.6.4
triton==2.3.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.8.5
uri-template==1.3.0
urllib3==2.3.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
